+++ Jenkinsfile
@@ -1,0 +1,12 @@
+pipeline {
+    agent any
+    stages {
+        stage('SCRIPT - OLÁ JENKINS') {
+            steps {
+                script {
+                    sh "echo BEM VINDO"
+                }
+            }
+        }
+    }
+}
