+++ src/main/java/br/com/phmiranda/oracle/basico/Programa.java
@@ -1,0 +1,24 @@
+/*
+ * Author: phmiranda
+ * Project: desenvolvedor-oracle-se
+ * User Story: HU001
+ * Description: CURSO DE DESENVOLVIMENTO JAVA SE - PARTE 01 (AULA 01)
+ * Date: 30/08/2021
+ */
+
+package br.com.phmiranda.oracle.basico;
+
+/*
+ * 1 - No Java, declaramos uma classe onde iremos desenvolver nosso código entre chaves (escopo) e colocamos o código a ser executado.
+ */
+public class Programa {
+    /*
+     * 2 - Aplicativos Java precisam invocar um método main (principal).
+     */
+    public static void main(String[] args) {
+        /*
+         * 3 - Método padrão para a impressão de uma string no console.
+         */
+        System.out.println("OLÁ MUNDO !!!");
+    }
+}
