+++ src/main/java/br/com/phmiranda/oracle/basico/fluxo/condicional/BaseCondicional.java
@@ -6,12 +6,11 @@
  * Date: 21/07/2021
  */
 
-package br.com.phmiranda.oracle.basico.condicional;
+package br.com.phmiranda.oracle.basico.fluxo.condicional;
 
 public class BaseCondicional {
     public static void main(String[] args) {
         int condicao = 1;
-
         if (condicao >= 0) {
             System.out.println("RESULTADO: " + condicao);
         } else {
