+++ src/main/java/br/com/phmiranda/oracle/basico/fluxo/controle/FluxoControleMultiplo.java
@@ -6,7 +6,7 @@
  * Date: 21/07/2021
  */
 
-package br.com.phmiranda.oracle.basico.controle;
+package br.com.phmiranda.oracle.basico.fluxo.controle;
 
 public class FluxoControleMultiplo {
     public static void main(String[] args) {
