+++ src/main/java/br/com/phmiranda/oracle/basico/primitivo/Byte.java
@@ -6,10 +6,11 @@
  * Date: 21/07/2021
  */
 
-package br.com.phmiranda.oracle.basico.controle;
+package br.com.phmiranda.oracle.basico.primitivo;
 
-public class BaseWhile {
+public class Byte {
     public static void main(String[] args) {
-
+        byte numero = 127;
+        System.out.println("Número em Byte: "+ numero);
     }
 }
