+++ src/main/java/br/com/phmiranda/oracle/basico/primitivo/Float.java
@@ -6,10 +6,11 @@
  * Date: 21/07/2021
  */
 
-package br.com.phmiranda.oracle.basico.controle;
+package br.com.phmiranda.oracle.basico.primitivo;
 
-public class BaseFor {
+public class Float {
     public static void main(String[] args) {
-
+        float numero = 3.14f;
+        System.out.println("Ponto Flutuante: " + numero);
     }
 }
