--- src/main/java/br/com/phmiranda/oracle/basico/sintaxe/Programa.java
@@ -1,19 +1,0 @@
-/*
- * Author: phmiranda
- * Project: desenvolvedor-oracle-se
- * User Story: HU001
- * Description: DESENVOLVEDOR ORACLE SE [PARTE 01]
- * Date: 21/07/2021
- */
-
-package br.com.phmiranda.oracle.basico.sintaxe;
-
-// No Java, declaramos uma classe onde iremos desenvolver nosso código entre chaves (escopo) e colocamos o código a ser executado.
-// Aplicativos Java precisam invocar um método main (principal).
-// Método padrão para a impressão de uma string no console.
-
-public class Programa {
-    public static void main(String[] args) {
-        System.out.println("Olá mundo");
-    }
-}
